refactor: extract navbar components in separate files
NavBar, NewPostPopup, NavButton...
